refactor tests structure
